requirements.txt
PyGitHub==1.51
requests~=2.27.1